feat: Add username prop to Home component and profile page
This commit adds a new prop, `username`, to the `Home` component in the `client/src/pages/Home.tsx` file. The prop is also passed down to the `Home` component in the `client/src/pages/Profile.tsx` file. This prop is used to navigate to the user's profile page when the "Go to Profile" button is clicked.